python/requirements.txt
pandas
mysql.connector